+++ test/kubernetes/excluster_test.go
@@ -1,11 +1,12 @@
-// +build k8s
+// +build ciskip
 
-package test
+package kubernetes
 
 import (
 	"testing"
 
 	"github.com/coredns/coredns/plugin/test"
+	intTest "github.com/coredns/coredns/test"
 
 	"github.com/miekg/dns"
 )
@@ -24,12 +25,10 @@ func TestKubernetesAPIFallthrough(t *testing.T) {
 	corefile :=
 		`.:0 {
     kubernetes cluster.local {
-        endpoint http://nonexistance:8080,http://invalidip:8080,http://localhost:8080
-        namespaces test-1
-        pods disabled
+        endpoint nonexistance:8080,invalidip:8080,localhost:8080
     }`
 
-	server, udp, _, err := CoreDNSServerAndPorts(corefile)
+	server, udp, _, err := intTest.CoreDNSServerAndPorts(corefile)
 	if err != nil {
 		t.Fatalf("Could not get CoreDNS serving instance: %s", err)
 	}
